requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
yfinance>=0.2.20
plotly>=5.15.0
numpy>=1.24.0
plaid-python>=12.0.0
requests>=2.31.0
python-dotenv>=1.0.0 